pyproject.toml
[project]
name = "llm-service-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "google-generativeai>=0.8.5",
    "openai>=1.76.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
]
